+++ src/ros2_medkit_gateway/src/http/handlers/data_handlers.cpp
@@ -25,6 +25,7 @@
 #include "ros2_medkit_gateway/core/plugins/plugin_manager.hpp"
 #include "ros2_medkit_gateway/core/providers/data_provider.hpp"
 #include "ros2_medkit_gateway/gateway_node.hpp"
+#include "ros2_medkit_serialization/type_introspection.hpp"
 
 using json = nlohmann::json;
